Introduce a pim directory to save the PIM index.
